+++ src/FirebaseConfig.js
@@ -33,3 +33,27 @@ export const provider = new GoogleAuthProvider();
 export const googlePopUp = (auth,provider) => {
   return signInWithPopup(auth, provider)
 };
+
+export const googleLogin = () => {
+  googlePopUp(auth, provider)
+      .then((result) => {
+        // This gives you a Google Access Token. You can use it to access the Google API.
+        const credential = GoogleAuthProvider.credentialFromResult(result);
+        const token = credential.accessToken;
+        // The signed-in user info.
+        const user = result.user;
+        // ...
+        window.location.assign('#/accesorios')
+        console.log('logueado con google')
+      }).catch((error) => {
+        // Handle Errors here.
+        const errorCode = error.code;
+        const errorMessage = error.message;
+        // The email of the user's account used.
+        const email = error.email;
+        // The AuthCredential type that was used.
+        const credential = GoogleAuthProvider.credentialFromError(error);
+        // ...
+      });
+      console.log('probando')
+}
